pyproject.toml
[project]
name = "openai-sdk-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hashir Mahmood", email = "hashir704x@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]

[project.scripts]
openai-sdk-work = "openai_sdk_work:main"
async-python = "openai_sdk_work.async_python:runMain"
simple-agent = "openai_sdk_work.simple_agent:runAgent"
output-type-agent = "openai_sdk_work.output_type:runAgent"
tool-calling-agent = "openai_sdk_work.tool_calling:runAgent"
handoffs-agent = "openai_sdk_work.handoffs:runAgent"
handoffs-2-agent = "openai_sdk_work.handoffs_2:runAgent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
